+++ inst/examples/Ex_math_support_opencl.R
@@ -1,5 +1,4 @@
-# See ?nmathopencl_examples_use_opencl
-if (nmathopencl_examples_use_opencl()) {
+if (!has_opencl() || identical(Sys.getenv("NOT_CRAN"), "true")) {
   imax2_opencl(x = 7, y = 3, fallback = FALSE, verbose = TRUE)
   imin2_opencl(x = 7, y = 3, fallback = FALSE, verbose = TRUE)
   fmax2_opencl(x = 7.2, y = 3.1, fallback = FALSE, verbose = TRUE)