+++ inst/examples/Ex_normal_opencl.R
@@ -1,7 +1,6 @@
-# See ?nmathopencl_examples_use_opencl
 n <- 5L
 x <- c(-1, 0, 1)
-if (nmathopencl_examples_use_opencl()) {
+if (!has_opencl() || identical(Sys.getenv("NOT_CRAN"), "true")) {
   dnorm_opencl(x, mean = 0, sd = 1, fallback = FALSE, verbose = TRUE)
   pnorm_opencl(q = 0.2, mean = 0, sd = 1, fallback = FALSE, verbose = TRUE)
   qnorm_opencl(rep(0.8, n), mean = 0, sd = 1, fallback = FALSE, verbose = TRUE)