+++ inst/examples/Ex_wilcox_opencl.R
@@ -1,5 +1,4 @@
-# See ?nmathopencl_examples_use_opencl
-if (nmathopencl_examples_use_opencl()) {
+if (!has_opencl() || identical(Sys.getenv("NOT_CRAN"), "true")) {
   # Wilcoxon OpenCL kernels are currently known to fail on some GPU stacks due
   # to unresolved runtime symbols. Keeping these commented avoids flaky checks:
   # n <- 1L